requirements.txt
pandas
scikit-learn
streamlit
plotly
seaborn
scipy
pillow